+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.2.1] - 2019-03-20
+### Fixed
+- Calling 'instance' API method shouldn't create an instance if it doesn't exist
+### Changed
+- Added documentation for 'instance' API method
+- Bumped year
+
 ## [1.2.0] - 2018-01-18
 ### Changed
 - Updated dependencies
@@ -12,5 +19,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - Call ```setOptions``` method on sub-sequent calls [\#1](https://github.com/kasparsz/jquery-plugin-generator/issues/1)
 
+[1.2.1]: https://github.com/kasparsz/jquery-plugin-generator/compare/v1.2.0...v1.2.1
 [1.2.0]: https://github.com/kasparsz/jquery-plugin-generator/compare/v1.1.0...v1.2.0
 [1.1.0]: https://github.com/kasparsz/jquery-plugin-generator/compare/v1.0.0...v1.1.0